requirements.txt
streamlit==1.26.0
openai==0.27.8
python-dotenv==1.0.0
langchain-openai==0.1.1
pandas==1.5.3
pandasql==0.7.3

